Refactor(#77): 모임 생성 시 동작 추가
