+++ src/components/create/CreateForm.tsx
@@ -1,12 +1,11 @@
 "use client";
 
 import { useMutation } from "@tanstack/react-query";
-import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { FormProvider, type SubmitHandler, useForm } from "react-hook-form";
-import SolidButton from "../common/buttons/SolidButton";
 import FormSectionLeft from "./FormSectionLeft";
 import FormSectionRight from "./FormSectionRight";
+import { FailModal, SuccessModal } from "./modals/ResultInfoModal";
 import useModal from "@/hooks/useModal";
 import { createMeetup } from "@/lib/main/meetup.api";
 import type { MeetupFormType } from "@/types/meetup.type";
@@ -28,73 +27,23 @@ export default function CreateForm() {
     },
     mode: "onChange",
   });
-
   const { control, handleSubmit, watch, setValue } = methods;
-
   const [image, setImage] = useState<File | null>(null);
-
-  const [dateError, setDateError] = useState<string | null>(null);
-
-  const { openModal, closeModal } = useModal();
-  const router = useRouter();
+  const { openModal } = useModal();
 
   const createMeetupMutation = useMutation({
     mutationFn: createMeetup,
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     onSuccess: (data: any) => {
-      console.log("Meetup created successfully:", data);
-
       openModal({
         hasCloseBtn: false,
-        children: (
-          <div className='flex w-[17.6875rem] flex-col items-center p-6'>
-            <p className='pb-3 text-heading-2 font-medium text-gray-100'>
-              모임 개설 완료
-            </p>
-            <p className='pb-6 text-body-2-normal font-medium text-gray-400'>
-              모임 개설이 완료되었어요
-            </p>
-            <div className='flex w-full gap-[.4375rem]'>
-              <SolidButton
-                onClick={() => {
-                  closeModal();
-                  router.push("/");
-                }}
-              >
-                목록으로
-              </SolidButton>
-              <SolidButton
-                onClick={() => {
-                  closeModal();
-                  router.push(`/study/${data.data.meetupId}`);
-                }}
-                state='activated'
-              >
-                보러가기
-              </SolidButton>
-            </div>
-          </div>
-        ),
+        children: <SuccessModal data={data} />,
       });
     },
-    onError: (error: Error) => {
-      console.error("Error creating meetup:", error);
-
+    onError: () => {
       openModal({
         hasCloseBtn: false,
-        children: (
-          <div className='flex w-[17.6875rem] flex-col items-center p-6'>
-            <p className='pb-3 text-heading-2 font-medium text-gray-100'>
-              모임 개설 실패패
-            </p>
-            <p className='pb-6 text-body-2-normal font-medium text-gray-400'>
-              모임 개설 중 오류가 발생했어요.
-            </p>
-            <div className='flex w-full gap-[.4375rem]'>
-              <SolidButton onClick={() => closeModal()}>닫기</SolidButton>
-            </div>
-          </div>
-        ),
+        children: <FailModal />,
       });
     },
   });
@@ -104,27 +53,10 @@ export default function CreateForm() {
       const formData = new FormData();
       formData.append("request", JSON.stringify(data));
       formData.append("image", image || "");
-      console.log("Meetup form data:", formData.get("request"));
-      console.log("Meetup image data:", formData.get("image"));
 
       await createMeetupMutation.mutateAsync(formData);
     } catch (error) {
       console.error("Error creating meetup:", error);
-    }
-  };
-
-  const validateDates = (
-    recruitmentEndDate: Date | null,
-    meetingStartDate: Date | null,
-  ) => {
-    if (
-      recruitmentEndDate &&
-      meetingStartDate &&
-      meetingStartDate < recruitmentEndDate
-    ) {
-      setDateError("진행 기간은 모집 기간보다 과거일 수 없습니다.");
-    } else {
-      setDateError(null);
     }
   };
 
@@ -154,8 +86,6 @@ export default function CreateForm() {
           control={control}
           methods={methods}
           isSubmitDisabled={isSubmitDisabled}
-          dateError={dateError || ""}
-          validateDates={validateDates}
         />
       </form>
     </FormProvider>
