+++ src/hooks/auths/useTokenState.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useEffect, useState } from "react";
 
 const getCookie = (name: string): string | null => {
